3_analysis/eyetracking/requirements.txt
matplotlib
minineedle
numpy
pandas
Pillow
researchpy
scipy
seaborn